pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-agent"
version = "0.1.0"
description = "An open-source AI agent that controls GitLab via natural language"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
gitlab-agent = "gitlab_agent.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.setuptools.packages.find]
include = ["gitlab_agent*"]
